Move schema to config
